joint angles are now published as doubles.  (previously published as floats...)

+++ src/org/myrobotlab/service/interfaces/IKJointAnglePublisher.java
@@ -7,6 +7,6 @@ public interface IKJointAnglePublisher {
 
 	public String getName();
 
-	public Map<String, Float> publishJointAngles(HashMap<String, Float> angleMap);
+	public Map<String, Double> publishJointAngles(HashMap<String, Double> angleMap);
 
 }
